+++ example/app/resources/products/user_resource.rb
@@ -1,7 +1,6 @@
 class Products::UserResource < Products::BaseResource
   accessible_attributes :description, :model, :price_in_cents, :vendor
-  
-  def resource_scope
+  repository :scope do
     accessor.products
   end
 end