+++ lib/cyrax/resource.rb
@@ -1,5 +1,6 @@
 class Cyrax::Resource < Cyrax::Base
   include Cyrax::Extensions::HasResource
+  include Cyrax::Extensions::HasRepository
   include Cyrax::Extensions::HasService
   include Cyrax::Extensions::HasDecorator
   include Cyrax::Extensions::HasSerializer