+++ spec/cyrax/resource_spec.rb
@@ -10,11 +10,7 @@ module Cyrax
         it{ should respond_to(:params) }
         it{ should respond_to(:resource_name) }
         it{ should respond_to(:resource_class) }
-        it{ should respond_to(:resource_scope) }
         it{ should respond_to(:resource_attributes) }
-        it{ should respond_to(:build_resource) }
-        it{ should respond_to(:build_collection) }
-        it{ should respond_to(:find_resource) }
         it{ should respond_to(:assign_resource) }
         it{ should respond_to(:respond_with) }
         it{ should respond_to(:set_message) }
@@ -25,17 +21,19 @@ module Cyrax
     end
 
     subject { Cyrax::Resource.new }
+    let(:repository) { double }
     let(:resource) { double.as_null_object }
     let(:collection) { [double] }
     before do
       subject.stub(:params).and_return({id:123})
-      subject.stub(:find_resource).and_return(resource)
+      subject.stub(:repository).and_return(repository)
+      repository.stub(:find).and_return(resource)
       subject.class.send :resource, :foo
     end
 
     describe '#read_all' do
       it 'responds with decorated collection' do
-        subject.should_receive(:build_collection).and_return(collection)
+        repository.should_receive(:find_all).and_return(collection)
         subject.should_receive(:respond_with).with(collection, name: 'foos', present: :collection)
         subject.read_all
       end
@@ -43,7 +41,7 @@ module Cyrax
 
     describe '#build' do
       it 'responds with resource' do
-        subject.should_receive(:build_resource).with(nil).and_return(resource)
+        repository.should_receive(:build).with(nil).and_return(resource)
         subject.should_receive(:respond_with).with(resource)
         subject.build
       end
@@ -51,7 +49,7 @@ module Cyrax
 
     describe '#edit' do
       it 'responds with resource' do
-        subject.should_receive(:find_resource)
+        repository.should_receive(:find)
         subject.should_receive(:respond_with).with(resource)
         subject.edit
       end
@@ -59,36 +57,33 @@ module Cyrax
 
     describe '#read' do
       it 'responds with resource' do
-        subject.should_receive(:find_resource)
+        repository.should_receive(:find)
         subject.should_receive(:respond_with).with(resource)
         subject.read
       end
     end
 
     describe '#destroy' do
-      it 'responds with resource' do
-        subject.should_receive(:find_resource)
+      it 'destroys resource and responds with resource' do
+        repository.should_receive(:find)
+        repository.should_receive(:delete)
         subject.should_receive(:respond_with).with(resource)
-        subject.destroy
-      end
-
-      it 'destroys resource' do
-        resource.should_receive(:destroy)
         subject.destroy
       end
     end
 
     describe '#create' do
       let(:params) { {foo: 'bar'} }
-      before { subject.stub(:build_resource).and_return(resource) }
+      before { repository.stub(:build).and_return(resource) }
       it 'responds with resource' do
-        subject.should_receive(:build_resource).with(nil, params)
+        repository.should_receive(:save).with(resource)
+        repository.should_receive(:build).with(nil, params)
         subject.should_receive(:respond_with).with(resource)
         subject.create(params)
       end
 
       context 'when resource successfully saved' do
-        before { resource.stub(:save).and_return(true) }
+        before { repository.stub(:save).and_return(true) }
 
         it 'sets message' do
           subject.should_receive(:set_message).with(:created)
@@ -97,7 +92,7 @@ module Cyrax
       end
 
       context 'when resource could not be saved' do
-        before { resource.stub(:save).and_return(false) }
+        before { repository.stub(:save).and_return(false) }
 
         it 'does not set message' do
           subject.should_not_receive(:set_message).with(:created)
@@ -108,15 +103,16 @@ module Cyrax
 
     describe '#update' do
       let(:params) { {foo: 'bar'} }
-      before { subject.stub(:build_resource).and_return(resource) }
+      before { repository.stub(:build).and_return(resource) }
       it 'responds with resource' do
-        subject.should_receive(:build_resource).with(123, params)
+        repository.should_receive(:build).with(123, params)
+        repository.should_receive(:save).with(resource)
         subject.should_receive(:respond_with).with(resource)
         subject.update(params)
       end
 
       context 'when resource successfully saved' do
-        before { resource.stub(:save).and_return(true) }
+        before { repository.stub(:save).and_return(true) }
 
         it 'sets message' do
           subject.should_receive(:set_message).with(:updated)
@@ -125,7 +121,7 @@ module Cyrax
       end
 
       context 'when resource could not be saved' do
-        before { resource.stub(:save).and_return(false) }
+        before { repository.stub(:save).and_return(false) }
 
         it 'does not set message' do
           subject.should_not_receive(:set_message).with(:created)